+++ recipes/default.rb
@@ -4,9 +4,6 @@
 #
 # Copyright:: 2018, The Authors, All Rights Reserved.
 
-package 'unzip' do
-  action :install
-end
-
+include_recipe 'build_agent::basic'
 include_recipe 'build_agent::packer'
 include_recipe 'build_agent::docker'