+++ spec/unit/recipes/default_spec.rb
@@ -7,25 +7,17 @@
 require 'spec_helper'
 
 describe 'build_agent::default' do
-  context 'When all attributes are default, on Ubuntu 16.04' do
+
+  context 'When all attributes are default, on CentOS 6.9' do
     let(:chef_run) do
-      # for a complete list of available platforms and versions see:
-      # https://github.com/customink/fauxhai/blob/master/PLATFORMS.md
-      runner = ChefSpec::ServerRunner.new(platform: 'ubuntu', version: '16.04')
+      runner = ChefSpec::ServerRunner.new(platform: 'centos', version: '6.9')
       runner.converge(described_recipe)
     end
 
-    it 'converges successfully' do
-      expect { chef_run }.to_not raise_error
-    end
-  end
-
-  context 'When all attributes are default, on CentOS 7.4.1708' do
-    let(:chef_run) do
-      # for a complete list of available platforms and versions see:
-      # https://github.com/customink/fauxhai/blob/master/PLATFORMS.md
-      runner = ChefSpec::ServerRunner.new(platform: 'centos', version: '7.4.1708')
-      runner.converge(described_recipe)
+    it 'Installs basic packages' do
+      expect(chef_run).to install_package('dos2unix')
+      expect(chef_run).to install_package('screen')
+      expect(chef_run).to install_package('unzip')
     end
 
     it 'converges successfully' do