+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # CHANGELOG
 
-## 0.1.0 Initial release
-
-## 0.1.1 Major rewrite of Packer and Docker parts
+- 0.1.0 Initial release
+- 0.1.1 Major rewrite of Packer and Docker parts
+- 0.1.2 Added installation of OS packages